requirements.txt
fastapi
uvicorn
pydantic
celery
redis
boto3
python-dotenv
loguru
requests